Implemented VST MIDI effect plugin. May not be necessary if the VST plugin can generate events that can be "frozen."

+++ frontends/CsoundVST/ScoreGeneratorVst.hpp
@@ -67,7 +67,7 @@ protected:
     {
       kNumPrograms = 10
     };
-  std::multimap<int, ScoreGeneratorEvent> scoreGeneratorEvents;
+  std::multimap<double, ScoreGeneratorEvent> scoreGeneratorEvents;
   std::vector<VstMidiEvent> vstMidiEventsBuffer;
   VstEvents *vstEventsPointer;
   double vstFramesPerSecond;
